Корректировки

+++ src/main/java/ru/skypro/homework/entity/Announce.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.stereotype.Repository;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -20,13 +19,13 @@ public class Announce {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long pk;
+    private int pk;
     private String authorFirstName;
     private String authorLastName;
     private String description;
     private String email;
     private String image;
     private String phone;
-    private Long price;
+    private int price;
     private String title;
 }
